+++ TwitterVoice/src/main/java/kai/twitter/voice/MainActivity.java
@@ -8,16 +8,17 @@ import android.os.Bundle;
 import android.support.v7.app.ActionBarActivity;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
 import android.widget.CompoundButton;
 
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.AdView;
 
+import org.jraf.android.backport.switchwidget.Switch;
+
 import kai.twitter.voice.manageAccount.ManageAccountsActivity;
 
-public class MainActivity extends ActionBarActivity implements View.OnClickListener {
-    private CompoundButton startServiceToggle;
+public class MainActivity extends ActionBarActivity implements CompoundButton.OnCheckedChangeListener {
+    private Switch startServiceToggle;
     private DbAdapter adapter;
     private AdView adView;
     private ServiceReceiver serviceReceiver;
@@ -29,8 +30,8 @@ public class MainActivity extends ActionBarActivity implements View.OnClickListe
 
         adapter = new DbAdapter(getApplicationContext());
 
-        startServiceToggle = (CompoundButton) findViewById(R.id.switch_start_service);
-        startServiceToggle.setOnClickListener(this);
+        startServiceToggle = (Switch) findViewById(R.id.switch_start_service);
+        startServiceToggle.setOnCheckedChangeListener(this);
 
         serviceReceiver = new ServiceReceiver();
         IntentFilter filter = new IntentFilter();
@@ -95,10 +96,10 @@ public class MainActivity extends ActionBarActivity implements View.OnClickListe
     }
 
     @Override
-    public void onClick(View view) {
+    public void onCheckedChanged(CompoundButton view, boolean isChecked) {
         if (startServiceToggle.getId() == view.getId()) {
             Intent serviceIntent = new Intent(getApplicationContext(), TwitterVoiceService.class);
-            if (startServiceToggle.isChecked()) {
+            if (isChecked) {
                 if (adapter.getAccounts().isEmpty()) {
                     Intent accountManageIntent = new Intent(getApplicationContext(), ManageAccountsActivity.class);
                     startActivity(accountManageIntent);