+++ changelog.md
@@ -3,6 +3,7 @@
 
 - Open source license notify for Switch backport.
 - Mute time works correctly.
+- Remove URL correctly.
 
 
 1.1.2