+++ changelog.md
@@ -2,6 +2,7 @@
 -----
 - ActionBar background color.
 - Switch works on slide.
+- Image button on main activity.
 
 1.1.3
 -----